refactoring AbstractPea, FreezePea, Pea. remove package name in SunProduct.java
