+++ src/FreezePea.java
@@ -9,25 +9,4 @@ public class FreezePea extends AbstractPea {
     public FreezePea(GamePanel parent, int lane, int startX) {
         super(parent, lane, startX);
     }
-
-    @Override
-    public void advance() {
-        plantMovingStrategy.move();
-    }
-
-    public void setMovingStrategy(PlantMovingStrategy movingStrategy) {
-        this.plantMovingStrategy = movingStrategy;
-    }
-
-    public int getPosX() {
-        return plantMovingStrategy.getPosX();
-    }
-
-    public void setPosX(int posX) {
-        plantMovingStrategy.setPosX(posX);
-    }
-    
-    protected void setPosXX() {
-    	setPosX(getPosX() - 15);
-    }
 }
